fix(wallet): emit internal error if something broke while broadcast
